+++ R/Gibbs_OMFA.R
@@ -34,12 +34,12 @@
       psi.store    <- array(0, dim=c(P, G, n.store))
     }
     if(sw["pi.sw"])  {
-      pi.store     <- matrix(0, nrow=G, ncol=n.store)
+      pi.store     <- matrix(0,  nrow=G, ncol=n.store)
     }
-    z.store        <- matrix(0, nrow=N, ncol=n.store)
-    ll.store       <- rep(0, n.store)
+    z.store        <- matrix(0L, nrow=N, ncol=n.store)
+    ll.store       <- rep(0,  n.store)
     err.z          <- zerr <- FALSE
-    G.store        <- ll.store
+    G.store        <- rep(0L, n.store)
 
     mu.sigma       <- 1/sigma.mu
     sig.mu.sqrt    <- sqrt(sigma.mu)
@@ -161,9 +161,9 @@
         if(all(sw["l.sw"], Q0))     load.store[,,,new.it]   <- lmat
         if(sw["psi.sw"])            psi.store[,,new.it]     <- 1/psi.inv
         if(sw["pi.sw"])             pi.store[,new.it]       <- pi.prop
-                                    z.store[,new.it]        <- z
+                                    z.store[,new.it]        <- as.integer(z)
                                     ll.store[new.it]        <- sum(z.res$log.like)
-                                    G.store[new.it]         <- G.non
+                                    G.store[new.it]         <- as.integer(G.non)
       }
     }
     if(verbose)       close(pb)